Add a new Rubocop that requires frozen_string_literal
The only existing Rubocop regarding frozen_string_literal is one that
simply requires the comment be present in _every_ file. The Ruby 3.4
behavior is described as requiring the frozen_string_literal comment in
files that define string literals. Add a new custom cop to replicate
that behavior, so we can run it on new files and work on cleaning up
existing non-frozen string literals.